Support keyboard shortcut for mark
